Initial attempt to externalize strings
In order to translate the webapp, the used strings should be somewhere
consolidated. There is now a `Messages.elm` file that is supposed to
contain all strings for the whole application. The language setting is
stored in browser's local storage. As an example, the login page has
been provided in German and English.