Use a configurable interval to refresh authentication token
Before a fixed interval of `session time - 30s` was used in the webapp
to re-authenticate in order to get a new authentication token. This
time may be too small if requests get stalled. It can now be
configured and the default is changed to a session time of 8 minutes
and a refresh time of 4 minutes. Iow a token is valid for 8 minutes
and the webapp will retrieve a new one every 4 minutes.

Refs: #132