Fix storing current chunk size
The chunks were saved to the database, but the size was only saved
after the complete request was processed. In case of a cancelled
request (due to timeouts for example) this led to wrongly advertising
of how much was already saved and client was uploading chunks again
although already at the server.

This makes uploading more reliable for larger request payloads where
timeouts may occur at the server cancelling the request.